[FIX/#22] CreateFragment 용도 변경
